@@ -1,9 +1,10 @@
 package com.futuremind.daggerutils
 
+import androidx.lifecycle.ViewModel
 import kotlin.annotation.AnnotationRetention.RUNTIME
 import kotlin.reflect.KClass
 
 @Target(AnnotationTarget.FUNCTION)
 @Retention(RUNTIME)
 @dagger.MapKey
-annotation class ViewModelKey(val value: KClass<out android.arch.lifecycle.ViewModel>)
+annotation class ViewModelKey(val value: KClass<out ViewModel>)
